core+loaders: share the eligibility/hash scan; drop dead code
The 'check eligibility before hashing' modpack-folder scan existed word-for-word
in all three bootstrap entry points (NeoForge fml4, fml10/fml11, Forge) - now one
EarlyServiceScan.eligibleJars() in core, along with the private isJar copies it
absorbs. Also: fml47 LazyModLocator's never-read LOGGER field removed, and the
fml4 EarlyServiceLayer class javadoc no longer says 'neo/forge' for a class that
is NeoForge-only.

SPI-coverage audit result recorded for the record: decompiling the exact loader
jars each generation runs on (fancymodloader 4.0.42/10.0.32/11.0.13, fmlloader
1.18.2/1.19.2/1.20.1) confirms every early-service SPI each loader discovers is
either hosted in place or deliberately force-copied (ImmediateWindowProvider
only); 1.19.2 shares fml47's reflection safely (identical SERVICES field).
